refactor(web): change FPA to S78
