+++ appeals/web/src/server/appeals/appeal-details/accordions/index.js
@@ -1,2 +1,2 @@
 export { generateAccordion as generateHasAccordion } from './has.js';
-export { generateAccordion as generateFpaAccordion } from './fpa.js';
+export { generateAccordion as generateS78Accordion } from './s78.js';
